+++ frontend/src/APIClients/CommonAPIClient.ts
@@ -1,129 +1,8 @@
 import AUTHENTICATED_USER_KEY from "../constants/AuthConstants";
-import {
-  GetLogRecordsReponse,
-  GetLogRecordCountResponse,
-} from "../types/LogRecordTypes";
 import { getLocalStorageObjProperty } from "../utils/LocalStorageUtils";
 import baseAPIClient from "./BaseAPIClient";
-
-import {
-  LogRecordFilters,
-  CountLogRecordFilters,
-} from "../components/common/types/Filters";
 import { Resident } from "../types/ResidentTypes";
 import { CountUsersResponse, GetUsersResponse } from "../types/UserTypes";
-
-const filterLogRecords = async ({
-  building = "",
-  employeeId = [],
-  attnTo = [],
-  dateRange = [],
-  tags = [],
-  flagged = false,
-  returnAll = false,
-  pageNumber,
-  resultsPerPage,
-}: LogRecordFilters): Promise<GetLogRecordsReponse> => {
-  try {
-    const bearerToken = `Bearer ${getLocalStorageObjProperty(
-      AUTHENTICATED_USER_KEY,
-      "accessToken",
-    )}`;
-    const { data } = await baseAPIClient.get(`/log_records`, {
-      params: {
-        filters: {
-          building,
-          employeeId,
-          attnTo,
-          dateRange,
-          tags,
-          flagged,
-        },
-        returnAll,
-        pageNumber,
-        resultsPerPage,
-      },
-      headers: { Authorization: bearerToken },
-    });
-    return data;
-  } catch (error) {
-    // TODO: more descriptive error / throw an exception potentially?
-    return null;
-  }
-};
-
-const countLogRecords = async ({
-  building = "",
-  employeeId = [],
-  attnTo = [],
-  dateRange = [],
-  tags = [],
-  flagged = false,
-}: CountLogRecordFilters): Promise<GetLogRecordCountResponse> => {
-  try {
-    const bearerToken = `Bearer ${getLocalStorageObjProperty(
-      AUTHENTICATED_USER_KEY,
-      "accessToken",
-    )}`;
-    const { data } = await baseAPIClient.get(`/log_records/count`, {
-      params: {
-        filters: {
-          building,
-          employeeId,
-          attnTo,
-          dateRange,
-          tags,
-          flagged,
-        },
-      },
-      headers: { Authorization: bearerToken },
-    });
-    return data;
-  } catch (error) {
-    // TODO: more descriptive error / throw an exception potentially?
-    return null;
-  }
-};
-
-const getUsers = async (
-  pageNumber: number,
-  resultsPerPage: number,
-): Promise<GetUsersResponse> => {
-  try {
-    const bearerToken = `Bearer ${getLocalStorageObjProperty(
-      AUTHENTICATED_USER_KEY,
-      "accessToken",
-    )}`;
-    const { data } = await baseAPIClient.get<GetUsersResponse>(`/users`, {
-      params: {
-        pageNumber,
-        resultsPerPage,
-      },
-      headers: { Authorization: bearerToken },
-    });
-    return data;
-  } catch (error) {
-    return null;
-  }
-};
-
-const countUsers = async (): Promise<CountUsersResponse> => {
-  try {
-    const bearerToken = `Bearer ${getLocalStorageObjProperty(
-      AUTHENTICATED_USER_KEY,
-      "accessToken",
-    )}`;
-    const { data } = await baseAPIClient.get<CountUsersResponse>(
-      `/users/count`,
-      {
-        headers: { Authorization: bearerToken },
-      },
-    );
-    return data;
-  } catch (error) {
-    return null;
-  }
-};
 
 const inviteUser = async (
   email: string,
@@ -194,10 +73,6 @@ const createResident = async ({
 };
 
 export default {
-  filterLogRecords,
-  countLogRecords,
-  getUsers,
-  countUsers,
   inviteUser,
   isUserInvited: getUserStatus,
   createResident,
